+++ laika/reports.py
@@ -2,6 +2,7 @@
 
 import os
 import imp
+import csv
 import pytz
 import json
 import datetime as dt
@@ -529,6 +530,94 @@ class AdwordsReport(BasicReport):
             )
             first_client_id_processed = True
         return result
+
+
+class GoogleAdsReport(BasicReport):
+    """
+    Sends a query to Google Ads search API and returns the results as a csv
+    file. Query can be passed as a string or in a separate file in query_file
+    parameter.
+
+    More about reporting queries in Google Ads API:
+    https://developers.google.com/google-ads/api/docs/reporting/overview
+
+    Internally uses google-ads library to make the request.
+    """
+
+    query = None
+    query_file = None
+    customer_id = None
+    header = None
+    fieldnames = None
+
+    def __init__(self, *args, **kwargs):
+        super(GoogleAdsReport, self).__init__(*args, **kwargs)
+
+        self.formatter = FilenameFormatter(*args)
+        self.creds = self.conf['profiles'][self.profile]['credentials']
+        if self.customer_id:
+            if not isinstance(self.customer_id, list):
+                self.customer_id = [self.customer_id]
+        else:
+            raise ValueError('You have to specify customer_id!')
+
+    def process(self):
+        from google.protobuf import json_format
+        from google.api_core import protobuf_helpers
+        from google.ads.googleads.client import GoogleAdsClient
+
+        if not self.query and self.query_file:
+            with open(self.query_file) as f:
+                logging.info('Reading query from %s', self.query_file)
+                self.query = f.read()
+
+        self.query = self.formatter.format(self.query)
+        logging.info('Formatted query: %s', self.query)
+
+        client = GoogleAdsClient.load_from_string(self.creds)
+        service = client.get_service('GoogleAdsService')
+
+        results, fieldnames = [], None
+        for customer_id in self.customer_id:
+            logging.info('Querying for customer: %s', customer_id)
+            stream = service.search_stream(query=self.query, customer_id=customer_id)
+
+            for batch in stream:
+                for msg in batch.results:
+                    # MessageToDict is useful because it brings Enum names instead
+                    # of values by default. preserving_proto_field_name flag is
+                    # needed in order for fields to match with paths in field_mask.
+                    row = json_format.MessageToDict(msg, preserving_proto_field_name=True)
+                    fieldnames = batch.field_mask.paths
+
+                    # Here we flatten the message dict and filter out fields that
+                    # aren't explicitly selected in the query (like resource names)
+                    clean_row = []
+                    for path in fieldnames:
+                        # Fieldnames that are reserved python keywords appear in
+                        # python objects with trailing underscore. So far i
+                        # only found "type_". TODO: find a cleaner way of doing this
+                        if path.endswith('.type'):
+                            path += '_'
+
+                        # MessageToDict will not include null metrics sometimes,
+                        # we use a value from the original message in this case
+                        _default = protobuf_helpers.get(msg, path)
+                        value = protobuf_helpers.get(row, path, default=_default)
+                        clean_row.append(value)
+
+                    results.append(clean_row)
+
+        # Return result as a csv file
+        data = six.StringIO()
+        if self.header:
+            data.write(self.formatter.format(self.header) + '\n')
+        if self.fieldnames:
+            fieldnames = self.fieldnames
+        csv_writer = csv.writer(data)
+        csv_writer.writerow(fieldnames)
+        csv_writer.writerows(results)
+        return data
 
 
 class FacebookInsightsReport(BasicReport):
@@ -1665,6 +1754,7 @@ class Config(dict):
         'redash': RedashReport,
         'bash': BashReport,
         'adwords': AdwordsReport,
+        'googleads': GoogleAdsReport,
         'facebook': FacebookInsightsReport,
         'trackeame': TrackeameReport,
         'drive': DownloadFromGoogleDrive,