+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
  - Updated Facebook Graph API version to v12.0
  - Add option to download Redash result as csv
+ - Added Google Ads report
 
 ## [1.4.0] - 2020-07-12